Delete session.py

+++ database/session.py
@@ -1,7 +1,8 @@
 from database.model import db
 from sqlalchemy.orm import sessionmaker
-from database.model.models import User, Tag, Article
+from database.model.models import User, Tag, Article, Website
 import types
+from datetime import datetime
 
 Session = sessionmaker(bind=db.engine)
 
@@ -71,10 +72,10 @@ class Create():
         if not user_exist:
             if nickname is None or "":
                 nickname = email
-                user = User(email=email, nickname=nickname, password=password)
-                session.add(user)
-                session.commit()
-                print "User added with success"
+            user = User(email=email, nickname=nickname, password=password)
+            session.add(user)
+            session.commit()
+            print "User added with success"
         else:
             session.rollback()
             session.close()
@@ -84,6 +85,28 @@ class Create():
     @staticmethod
     def websites():
         pass
+
+    @staticmethod
+    def website(uri, name):
+        '''Create website.
+
+        Name and uri are necessary.
+        '''
+        if uri is None or "":
+            raise FunctionParameterError("Uri is necessary")
+        if name is None or "":
+            raise FunctionParameterError("Name is necessary")
+        session = Session()
+        website_exists = session.query(Website).filter_by(uri=uri).all()
+        if not website_exists:
+            session.add(Website(uri=uri, name=name))
+            session.commit()
+            print "Website added with success"
+        else:
+            session.rollback()
+            session.close()
+            raise AlreadyExists("Websie")
+        session.close()
 
     @staticmethod
     def tags(list_of_tags):
@@ -105,8 +128,8 @@ class Create():
         if name is None or "":
             raise FunctionParameterError("Name is necessary")
         session = Session()
-        tag_exist = session.query(Tag).filter_by(name=name).all()
-        if not tag_exist:
+        tag_exists = session.query(Tag).filter_by(name=name).all()
+        if not tag_exists:
             session.add(Tag(name=name))
             session.commit()
             print "Tag added with success"
@@ -114,32 +137,83 @@ class Create():
             session.rollback()
             session.close()
             raise AlreadyExists("Tag")
+        session.close()
 
     @staticmethod
     def articles():
         pass
 
+    @staticmethod
+    def article(head, content, time=None, picture=None):
+        '''Create article.
+
+        Head and content are necessary.
+        '''
+        if head is None or "":
+            raise FunctionParameterError("Head is necessary")
+        if content is None or "":
+            raise FunctionParameterError("Content is necessary")
+        session = Session()
+        if time is None:
+            time = datetime.utcnow()
+        session.add(Article(head=head, content=content, time=time,
+                            picture=picture))
+        session.commit()
+        session.close()
+
 
 class Delete():
-    def users(self):
-        pass
+    @staticmethod
+    def user(email):
+        session = Session()
+        user = session.query(User).filter_by(email=email).first()
+        session.delete(user)
+        session.commit()
+        session.close()
 
-    def tags(self):
-        pass
+    @staticmethod
+    def tag(name):
+        session = Session()
+        tag = session.query(Tag).filter_by(name=name).first()
+        session.delete(tag)
+        session.commit()
+        session.close()
 
-    def websites(self):
-        pass
+    @staticmethod
+    def website(uri):
+        session = Session()
+        website = session.query(Website).filter_by(uri=uri).first()
+        session.delete(website)
+        session.commit()
+        session.close()
 
-    def artticles(self):
-        pass
+    @staticmethod
+    def artticle(id_):
+        session = Session()
+        article = session.query(Article).filter_by(id=id_).first()
+        session.delete(article)
+        session.commit()
+        session.close()
 
     @staticmethod
     def all_tags():
-        pass
+        '''delete all tags'''
+        session = Session()
+        tags = session.query(Tag).all()
+        for t in tags:
+            session.delete(t)
+        session.commit()
+        session.close()
 
     @staticmethod
     def all_websties():
-        pass
+        '''delete all websites'''
+        session = Session()
+        websites = session.query(Website).all()
+        for w in websites:
+            session.delete(w)
+        session.commit()
+        session.close()
 
     @staticmethod
     def all_articles():
@@ -156,9 +230,10 @@ class Delete():
         '''Delete all users.'''
         session = Session()
         users = session.query(User).all()
-        for u in users:
-            session.delete(u)
-        session.commit()
+        if users:
+            for u in users:
+                session.delete(u)
+                session.commit()
         session.close()
 
 
@@ -170,3 +245,125 @@ class Get():
         users = session.query(User).all()
         session.close()
         return users
+
+    @staticmethod
+    def user(email):
+        '''Get user with given email'''
+        session = Session()
+        user = session.query(User).filter_by(email=email).first()
+        session.close()
+        return user
+
+    @staticmethod
+    def all_websites():
+        '''Get all websites as a list.'''
+        session = Session()
+        websites = session.query(Website).all()
+        session.close()
+        return websites
+
+    @staticmethod
+    def website(uri):
+        '''Get website with given uri'''
+        session = Session()
+        website = session.query(Website).filter_by(uri=uri).first()
+        session.close()
+        return website
+
+    @staticmethod
+    def all_tags():
+        '''Get all tags as a list.'''
+        session = Session()
+        tags = session.query(Tag).all()
+        session.close()
+        return tags
+
+    @staticmethod
+    def tag(name):
+        '''Get tag with given name'''
+        session = Session()
+        tag = session.query(Tag).filter_by(name=name).first()
+        session.close()
+        return tag
+
+    @staticmethod
+    def all_articles():
+        '''Get all articles as a list.'''
+        session = Session()
+        articles = session.query(Article).all()
+        session.close()
+        return articles
+
+    @staticmethod
+    def article(id_):
+        '''Get user with given email'''
+        session = Session()
+        article = session.query(Article).filter_by(id=id_).first()
+        session.close()
+        return article
+
+    @staticmethod
+    def users_tags(email):
+        '''Get user's tags as a list.'''
+        session = Session()
+        user = session.query(User).filter_by(email=email).first()
+        tags = user.tags.all()
+        session.close()
+        return tags
+
+    @staticmethod
+    def users_websites(email):
+        '''Get user's websites as a list'''
+        session = Session()
+        user = session.query(User).filter_by(email=email).first()
+        websites = []
+        tags = user.tags.all()
+        for tag in tags:
+            websites.append(tag.websites.all())
+        session.close()
+        return websites
+
+    @staticmethod
+    def tags_websites(name):
+        '''Get tag's websites as a list'''
+        session = Session()
+        tag = session.query(Tag).filter_by(name=name).first()
+        websites = tag.websites.all()
+        session.close()
+        return websites
+
+
+class Add():
+    @staticmethod
+    def tag_to_user(email, tagname):
+        '''Add tag to user'''
+        session = Session()
+        user_exists = session.query(User).filter_by(email=email).first()
+        tag_exists = session.query(Tag).filter_by(name=tagname).first()
+        if tag_exists and user_exists:
+            user_exists.tags.append(tag_exists)
+        session.commit()
+        session.close()
+
+    @staticmethod
+    def website_to_tag(tagname, website_uri):
+        '''Add website to tag'''
+        session = Session()
+        tag_exists = session.query(Tag).filter_by(name=tagname).first()
+        website_exists = session.query(Website).\
+            filter_by(uri=website_uri).first()
+        if tag_exists and website_exists:
+            tag_exists.websites.append(website_exists)
+        session.commit()
+        session.close()
+
+    @staticmethod
+    def article_to_website(website_uri, article_id):
+        '''Add article to website'''
+        session = Session()
+        website_exists = session.query(Website).filter_by(uri=website_uri).\
+            first()
+        article_exists = session.query(Article).filter_by(id=article_id).\
+            first()
+        if website_exists and article_exists:
+            website_exists.articles.append(article_exists)
